refactor: shared methods and properties for Badi components
