+++ component/calendar/mixins/template/Calendar.js
@@ -1,6 +1,5 @@
 import dashHas from 'lodash.has'
 import { DateTime } from 'luxon'
-import BadiDate from 'utils/badidate'
 
 const debug = require('debug')('calendar:Calendar')
 
@@ -40,9 +39,6 @@ export default {
   computed: {
     workingDateTime () {
       return this.makeDT(this.workingDate)
-    },
-    workingDateBadi: function () {
-      return new BadiDate(this.workingDateTime)
     }
   },
   methods: {
